Lines of code to patch issues with neg isolevels
Representation of iso-surfaces with semitransparence with negative
isovalues is not working properly. Although this will be probably
solved in futher nglview releases, it could be solved by now
changing the sign in the cube files. This would require generating
2 components (one for each cube file) for each MO. Let's keep this
piece of code, but not integrate this into the notebook for the
moment.

Another possibility, which may be more convinient is to add an option
to change the sign of the file.